requirements.txt
streamlit>=1.32.0
supabase>=2.3.0
python-dotenv>=1.0.0
